requirements.txt
icrawler==0.6.6
Pillow==10.4.0
requests==2.32.3
